Revert "Merge branch 'Development/Feature/ckw' into 'master'"
This reverts merge request !33

+++ src/main/java/nl/fontys/s3/huister/business/interfaces/city/GetAllCitiesUseCase.java
@@ -3,5 +3,5 @@ package nl.fontys.s3.huister.business.interfaces.city;
 import nl.fontys.s3.huister.business.response.city.GetAllCitiesResponse;
 
 public interface GetAllCitiesUseCase {
-    GetAllCitiesResponse getAllCities(long userId);
+    GetAllCitiesResponse getAllCities(int userId);
 }
