+++ src/main/java/nl/fontys/s3/huister/business/interfaces/property/GetAllPropertiesUseCase.java
@@ -5,5 +5,5 @@ import nl.fontys.s3.huister.business.response.property.GetAllPropertiesResponse;
 import java.util.List;
 
 public interface GetAllPropertiesUseCase {
-    List<GetAllPropertiesResponse>getAllProperties(final long userId);
+    List<GetAllPropertiesResponse>getAllProperties(final int userId);
 }
