+++ src/main/java/nl/fontys/s3/huister/business/interfaces/property/GetPropertyDetailUseCase.java
@@ -3,5 +3,5 @@ package nl.fontys.s3.huister.business.interfaces.property;
 import nl.fontys.s3.huister.business.response.property.GetPropertyDetailResponse;
 
 public interface GetPropertyDetailUseCase {
-    GetPropertyDetailResponse getPropertyDetail(final long id);
+    GetPropertyDetailResponse getPropertyDetail(final int id);
 }
