+++ src/main/java/nl/fontys/s3/huister/business/interfaces/user/GetUserByIdUseCase.java
@@ -4,5 +4,5 @@ package nl.fontys.s3.huister.business.interfaces.user;
 import nl.fontys.s3.huister.business.response.user.GetUserByIdResponse;
 
 public interface GetUserByIdUseCase {
-    GetUserByIdResponse getUserById(final long id);
+    GetUserByIdResponse getUserById(final int id);
 }
